FSP for H Bridge pulse ok, other not test yet

+++ Application/h_bridge_task/h_bridge_sequence.h
@@ -30,12 +30,12 @@ typedef struct _H_Bridge_task_typedef_
     uint8_t     lv_pos_count;
     uint8_t     lv_neg_count;
 
-    uint8_t     lv_delay_ms;
+    uint16_t    lv_delay_ms;
 
-    uint8_t     lv_pos_on_ms;
-    uint8_t     lv_pos_off_ms;
-    uint8_t     lv_neg_on_ms;
-    uint8_t     lv_neg_off_ms;
+    uint16_t    lv_pos_on_ms;
+    uint16_t    lv_pos_off_ms;
+    uint16_t    lv_neg_on_ms;
+    uint16_t    lv_neg_off_ms;
 } H_Bridge_task_typedef;
 
 /* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Variables ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ */
